added: vertex and index offset data to uniform buffer

+++ src/shaders/basic.vert
@@ -17,8 +17,8 @@ layout(binding = 0, set = 0) uniform Camera {
 
 layout(binding = 1, set = 0) uniform InstanceDescriptionContainer {
   uint instanceCount;
-  float vertexOffsets[8];
-  float indexOffsets[8];
+  uint vertexOffsets[8];
+  uint indexOffsets[8];
   mat4 transformMatrix[16];
 } instanceDescriptionContainer;
